requirements.txt
asgiref==3.3.4
Django==3.2.3
django-nested-admin==3.3.3
flake8==3.9.2
mccabe==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-monkey-business==1.0.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
